Cleanup Idea or Sonar warnings
Throw or log exception
Remove unused imports
Remove unused code
Unneeded parenthesis
Chain assertions in tests
...

+++ server/src/main/java/com/adobe/testing/s3mock/store/ObjectStore.java
@@ -54,8 +54,6 @@ import java.util.UUID;
 import java.util.concurrent.ConcurrentHashMap;
 import javax.xml.stream.XMLStreamException;
 import org.apache.commons.io.FileUtils;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 /**
  * Stores objects and their metadata created in S3Mock.
@@ -65,10 +63,9 @@ public class ObjectStore {
   private static final String ACL_FILE = "objectAcl.xml";
   private static final String DATA_FILE = "binaryData";
 
-  private static final Logger LOG = LoggerFactory.getLogger(ObjectStore.class);
   /**
    * This map stores one lock object per S3Object ID.
-   * Any method modifying the underlying file must aquire the lock object before the modification.
+   * Any method modifying the underlying file must acquire the lock object before the modification.
    */
   private final Map<UUID, Object> lockStore = new ConcurrentHashMap<>();
 
@@ -343,7 +340,6 @@ public class ObjectStore {
             sourceObject.checksum(),
             sourceObject.owner());
       } catch (IOException e) {
-        LOG.error("Can't write file to disk!", e);
         throw new IllegalStateException("Can't write file to disk!", e);
       }
     }
@@ -402,7 +398,6 @@ public class ObjectStore {
         try {
           FileUtils.deleteDirectory(getObjectFolderPath(bucket, id).toFile());
         } catch (IOException e) {
-          LOG.error("Can't delete directory.", e);
           throw new IllegalStateException("Can't delete directory.", e);
         }
         lockStore.remove(id);
@@ -432,10 +427,8 @@ public class ObjectStore {
   File inputStreamToFile(InputStream inputStream, Path filePath) {
     File targetFile = filePath.toFile();
     try {
-      if (targetFile.createNewFile()) {
-        if (!retainFilesOnExit) {
-          targetFile.deleteOnExit();
-        }
+      if (targetFile.createNewFile() && (!retainFilesOnExit)) {
+        targetFile.deleteOnExit();
       }
 
       try (InputStream is = inputStream;
@@ -448,7 +441,6 @@ public class ObjectStore {
         }
       }
     } catch (IOException e) {
-      LOG.error("Can't write file to disk!", e);
       throw new IllegalStateException("Can't write file to disk!", e);
     }
     return targetFile;
@@ -507,7 +499,6 @@ public class ObjectStore {
         return true;
       }
     } catch (IOException e) {
-      LOG.error("Could not write object metadata-file.", e);
       throw new IllegalStateException("Could not write object metadata-file.", e);
     }
   }
@@ -523,7 +514,6 @@ public class ObjectStore {
         return deserializeJaxb(toDeserialize);
       }
     } catch (IOException | JAXBException | XMLStreamException e) {
-      LOG.error("Could not write object metadata-file.", e);
       throw new IllegalStateException("Could not write object metadata-file.", e);
     }
   }
@@ -539,7 +529,6 @@ public class ObjectStore {
         return true;
       }
     } catch (IOException | JAXBException e) {
-      LOG.error("Could not write object metadata-file.", e);
       throw new IllegalStateException("Could not write object metadata-file.", e);
     }
   }
